requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
openai==1.101.0
